split additional.c
